move game log commands to own cog
